application/API/requirements.txt
fastapi[all]==0.80.0
fastapi-restful==0.4.3
fastapi-health==0.4.0
service-streamer==0.1.2
pydantic==1.9.2
numpy==1.23.3
onnxruntime==1.12.1
Pillow==9.2.0
loguru==0.6.0